+++ src/main/java/com/jeffpeng/jmod/actions/rotarycraft/AddBlastFurnaceAlloying.java
@@ -58,7 +58,7 @@ public class AddBlastFurnaceAlloying extends BasicAction {
 			return false;
 		}
 		valid = true;
-		execute();
+		//execute();
 		return valid;
 	}