+++ src/main/java/com/jeffpeng/jmod/scripting/mods/RotaryCraft.java
@@ -5,8 +5,11 @@ import com.jeffpeng.jmod.actions.rotarycraft.AddBlastFurnaceAlloying;
 import com.jeffpeng.jmod.actions.rotarycraft.AddBlastFurnaceRecipe;
 import com.jeffpeng.jmod.actions.rotarycraft.AddCentrifugeRecipe;
 import com.jeffpeng.jmod.actions.rotarycraft.AddCompactorRecipe;
+import com.jeffpeng.jmod.actions.rotarycraft.AddDryingBedRecipe;
 import com.jeffpeng.jmod.actions.rotarycraft.AddGrinderRecipe;
 import com.jeffpeng.jmod.actions.rotarycraft.AddPulseJetRecipe;
+import com.jeffpeng.jmod.actions.rotarycraft.AddRockMelterRecipe;
+import com.jeffpeng.jmod.actions.rotarycraft.AddWetterRecipe;
 import com.jeffpeng.jmod.primitives.OwnedObject;
 
 public class RotaryCraft extends OwnedObject {
@@ -40,6 +43,18 @@ public class RotaryCraft extends OwnedObject {
 	
 	public void addCompactorRecipe(String out, String in, int temperature, int pressure){
 		if(owner.testForMod("RotaryCraft")) new AddCompactorRecipe(owner,out,in,temperature, pressure);
+	}
+	
+	public void addDryingBedRecipe(String out, String in){
+		if(owner.testForMod("RotaryCraft")) new AddDryingBedRecipe(owner,out,in);
+	}
+	
+	public void addRockMelterRecipe(String out, String in, int temperature, long power){
+		if(owner.testForMod("RotaryCraft")) new AddRockMelterRecipe(owner,out,in,temperature,power);
+	}
+	
+	public void addLiquefactionRecipe(String out, String in, String fluidin, int time){
+		if(owner.testForMod("RotaryCraft")) new AddWetterRecipe(owner, out, in, fluidin, time);
 	}
 	
 	public void patchRotarycraftSteelTools(boolean value){